+++ src/index.js
@@ -1,4 +1,5 @@
 export { loadChangeInput } from './change-input.js';
+export { createChangeFromLocalDiff } from './local-diff.js';
 export { loadFrameworkRegistry, selectFramework } from './framework-registry.js';
 export { assessRisks, qualityConfidence } from './risk-engine.js';
 export { buildStrategy } from './strategy.js';